The architecture of that software its done like this: 1. there is a queue that its filled by `io.aiven.connect.elasticsearch.ElasticsearchSinkTask.put` 2. there is a task that takes a chunk of messages at a time and attempt to write those in ElasticSearch `io.aiven.connect.elasticsearch.ElasticsearchWriter.write`
Unfortunately, seems that if we reach the max retries for a record (or the record its not retriable).

We fill the atomic reference `io.aiven.connect.elasticsearch.bulk.BulkProcessor.error` without never freeing that variable.
Because of that, since each method its guarded by `io.aiven.connect.elasticsearch.bulk.BulkProcessor.throwIfTerminal`, like for e.g. the `io.aiven.connect.elasticsearch.bulk.BulkProcessor.add` method; isn't possible to forward once we fill this exception.

When the `dropInvalidMessage` its set to `true` in the `io.aiven.connect.elasticsearch.ElasticsearchWriter.tryWriteRecord` method (by setting the `drop.invalid.message` config to `"true"`) the try catches also the `bulkProcessor.add(record, flushTimeoutMs);` instruction while it should catch only the `convertRecord` method causing the connector to continuously dropping the new messages instead of dying, the semantics should be "I try to convert the record and if the shape it's not the expected one I skip the record", while the code was doing: "I try to convert and send async the messages in the queue and if any error happens I throw an exception and I skip the records".

This pr fixes it and add a test that expects an exception is thrown from the method when trying to add new records after an error happens in the OS client.